requirements.txt
django-bootstrap-toolkit
pydicom